v2.6
Finished adding in file system commands.

Added in networking commands.

Added in System commands.

Added steal_token command in gunnershell this steals the token of a process then spawns a shell as the owner of that token and steal_token integrates with the payload generator so you can  specify the rev shell payload you'd like + it can be fetched over HTTP or HTTPS.